1.4.3
materialowka / ilosci

+++ application/modules/services/controllers/ServicesController.php
@@ -386,6 +386,9 @@ class Services_ServicesController extends Application_Controller_Abstract {
             foreach ($defaults['installationcodeid'] as $i => $item) {
                 $defaults['installationcodeid-' . $i] = $item;
             }
+            foreach ($defaults['productid'] as $i => $item) {
+                $defaults['productid-' . $i] = $item;
+            }
         } else {
             $defaults = array('warehouseid' => $warehouseid,
                 'clientid' => $clientid,
@@ -397,18 +400,22 @@ class Services_ServicesController extends Application_Controller_Abstract {
         switch ($typeid) {
             // zlecenie instalacyjne
             case $types->find('installation', 'acronym')->id:
-                $form = new Application_Form_Services_Installation(array('installationCodesCount' => sizeof($defaults['installationcodeid'])));
+                $form = new Application_Form_Services_Installation(array('installationCodesCount' => sizeof($defaults['installationcodeid']),
+                    'productsCount' => sizeof($defaults['productid'])));
                 break;
             // zlecenie serwisowe
             case $types->find('service', 'acronym')->id:
-                $form = new Application_Form_Services_Service(array('installationCodesCount' => sizeof($defaults['installationcodeid'])));
+                $form = new Application_Form_Services_Service(array('installationCodesCount' => sizeof($defaults['installationcodeid']),
+                    'productsCount' => sizeof($defaults['productid'])));
                 break;
         }
         if ($service) {
             $form->setDefaults($service->toArray());
         }
-        $form->setOptions($options); //var_dump($options['installationcodes']);exit;
-        $form->setDefaults($defaults); //var_dump($defaults);exit;
+        $units = $this->_dictionaries->getDictionaryList('warehouse', 'unit');
+        $options['units'] = $units;
+        $form->setOptions($options); 
+        $form->setDefaults($defaults); 
         if (!in_array($this->_auth->getIdentity()->role, array('admin', 'coordinator'))) {
             $form->removeElement('coordinatorcomments');
         }
@@ -493,30 +500,59 @@ class Services_ServicesController extends Application_Controller_Abstract {
                 }
                 $service->save();
                 $status = $this->_dictionaries->getStatusList('orders')->find('invoiced', 'acronym');
+                $statusReleased = $this->_dictionaries->getStatusList('orders')->find('released', 'acronym');
+                $this->_orderlines->setLazyLoading(true);
                 foreach ($service->getProducts() as $product) {
+                    if ($orderLine = $this->_orderlines->find($product->productid)->current()) {
+                        $orderLine->statusid = $statusReleased->id;
+                        $orderLine->serviceid = null;
+                        $orderLine->qtyavailable += $orderLine->quantity;
+                        if ($orderLine->quantity > $orderLine->qtyavailable) {
+                            $form->getElement('productid')->setErrors(array('productid' => 'Wystąpił problem z dodaniem produktu'));
+                            return;
+                        }
+                        $orderLine->save();
+                    }
                     $product->delete();
                 }
                 $productIds = array_filter((array)$request->getParam('productid'));
                 if (!empty($productIds)) {
+                    $quantities = array_filter((array)$request->getParam('quantity'));
                     $table = new Application_Model_Services_Products_Table();
                     foreach ($productIds as $ix => $orderLineId) {
+                        $orderLineId = current($orderLineId);
+                        preg_match("/\d+/", $ix, $found);
+                        $ix = $found[0];
                         $product = $products->find($orderLineId);
-                        $params = array('serviceid' => $service->id);
+                        if (!($quantity = $quantities['quantity-' . $ix])) {
+                            $form->getElement('quantity-' . $ix)->setErrors(array('quantity-' . $ix => 'Brak ilości'));
+                            return;
+                        }
+                        $params = array('serviceid' => $service->id, 'quantity' => $quantity);
                         if (!$product) {
                             $id = $ix + 1;
                             $params['productid'] = "-$id";
                             $params['productname'] = $orderLineId;
-                            $form->getElement('productid')->addMultiOption($orderLineId, $orderLineId);
+                            $form->getElement('productid-' . $ix)->addMultiOption($orderLineId, $orderLineId);
                             //continue;
                         } else {
                             $params['productid'] = $product->id;
                             $params['productname'] = $product->getProduct()->name;
+                            $orderLine = $this->_orderlines->find($orderLineId)->current();
+                            $orderLine->statusid = $status->id;
+                            $orderLine->qtyavailable -= $quantity;//$orderLine->quantity;
+                            if ($orderLine->qtyavailable < 0) {
+                                $form->getElement('quantity-' . $ix)->setErrors(array('quantity-' . $ix => 'Wystąpił problem z dodaniem produktu'));
+                                return;
+                            }
+                            $orderLine->save();
                         }
+                        //var_dump($orderLineId,$params);exit;
                         $serviceProduct = $table->createRow($params);
                         $serviceProduct->save();
-                        $this->_orderlines->update(array('statusid' => $status->id, 'serviceid' => $service->id), $this->_orderlines->getAdapter()->quoteInto('id = ?', $orderLineId));
-                    }
-                }
+                        //$this->_orderlines->update(array('statusid' => $status->id, 'serviceid' => $service->id), $this->_orderlines->getAdapter()->quoteInto('id = ?', $orderLineId));
+                    }//exit;
+                }//return;
                 foreach ($service->getCodes() as $attribute) {//var_dump($attribute);continue;
                     $attribute->delete();
                 }
@@ -831,6 +867,7 @@ class Services_ServicesController extends Application_Controller_Abstract {
                                         $this->_users->setWhere($this->_users->getAdapter()->quoteInto("UPPER(lastname) = UPPER(?)", "{$data['lastname']}"));
                                         $user = $this->_users->getAll()->current();
                                     } else if (!empty($data['symbol'])) {
+                                        $this->_users->clearWhere();
                                         //$user = $this->_users->getAll(array('symbol' => $data['symbol']))->current();
                                         $this->_users->setWhere($this->_users->getAdapter()->quoteInto("UPPER(symbol) = UPPER(?)", "{$data['symbol']}"));
                                         $user = $this->_users->getAll()->current();
@@ -1117,7 +1154,7 @@ class Services_ServicesController extends Application_Controller_Abstract {
         $statusInvoiced = $this->_dictionaries->getStatusList('orders')->find('invoiced', 'acronym');
         $statusReleased = $this->_dictionaries->getStatusList('orders')->find('released', 'acronym');
         if ($service->technicianid) {
-            $this->_orderlines->setWhere("technicianid = '{$service->technicianid}' AND (statusid = {$statusReleased->id} OR (statusid = {$statusInvoiced->id} AND serviceid = {$service->id}))");
+            $this->_orderlines->setWhere("technicianid = '{$service->technicianid}' AND (statusid = {$statusReleased->id} OR (statusid = {$statusInvoiced->id} /*AND serviceid = {$service->id}*/))");
             $products = $this->_orderlines->getAll(
                 //array(
                 //'statusid' => array(
@@ -1261,7 +1298,19 @@ class Services_ServicesController extends Application_Controller_Abstract {
                 }
                 $service->save();
                 $status = $this->_dictionaries->getStatusList('orders')->find('invoiced', 'acronym');
+                $statusReleased = $this->_dictionaries->getStatusList('orders')->find('released', 'acronym');
+                $this->_orderlines->setLazyLoading(true);
                 foreach ($service->getProducts() as $product) {
+                    if ($orderLine = $this->_orderlines->find($product->productid)->current()) {
+                        $orderLine->statusid = $statusReleased->id;
+                        $orderLine->serviceid = null;
+                        $orderLine->qtyavailable += $orderLine->quantity;
+                        if ($orderLine->quantity > $orderLine->qtyavailable) {
+                            $form->getElement('productid')->setErrors(array('productid' => 'Wystąpił problem z dodaniem produktu'));
+                            return;
+                        }
+                        $orderLine->save();
+                    }
                     $product->delete();
                 }
                 $productIds = array_filter($request->getParam('productid'));
@@ -1278,6 +1327,14 @@ class Services_ServicesController extends Application_Controller_Abstract {
                         } else {
                             $params['productid'] = $product->id;
                             $params['productname'] = $product->getProduct()->name;
+                            $orderLine = $this->_orderlines->find($orderLineId)->current();
+                            $orderLine->statusid = $status->id;
+                            $orderLine->qtyavailable -= $orderLine->quantity;
+                            if ($orderLine->qtyavailable < 0) {
+                                $form->getElement('productid')->setErrors(array('productid' => 'Wystąpił problem z dodaniem produktu'));
+                                return;
+                            }
+                            $orderLine->save();
                         }
                         $serviceProduct = $table->createRow($params);
                         //$serviceProduct->serviceid = $service->id;
@@ -1286,7 +1343,7 @@ class Services_ServicesController extends Application_Controller_Abstract {
                             $serviceProduct->save();
                             
                         //}catch(Exception $e){var_dump($orderLineId,$e->getMessage(),$serviceProduct->toArray());exit;}
-                        $this->_orderlines->update(array('statusid' => $status->id, 'serviceid' => $service->id), $this->_orderlines->getAdapter()->quoteInto('id = ?', $orderLineId));
+                        //$this->_orderlines->update(array('statusid' => $status->id, 'serviceid' => $service->id), $this->_orderlines->getAdapter()->quoteInto('id = ?', $orderLineId));
                     }
                 }
                 foreach ($service->getCodes() as $attribute) {
@@ -1456,7 +1513,7 @@ class Services_ServicesController extends Application_Controller_Abstract {
         $statusInvoiced = $this->_dictionaries->getStatusList('orders')->find('invoiced', 'acronym');
         $statusReleased = $this->_dictionaries->getStatusList('orders')->find('released', 'acronym');
         if ($service->technicianid) {
-            $this->_orderlines->setWhere("technicianid = '{$service->technicianid}' AND (statusid = {$statusReleased->id} OR (statusid = {$statusInvoiced->id} AND serviceid = {$service->id}))");
+            $this->_orderlines->setWhere("technicianid = '{$service->technicianid}' AND (statusid = {$statusReleased->id} OR (statusid = {$statusInvoiced->id} /*AND serviceid = {$service->id}*/))");
             $products = $this->_orderlines->getAll(
                 //array(
                 //'statusid' => array(
@@ -1613,7 +1670,19 @@ class Services_ServicesController extends Application_Controller_Abstract {
                 }
                 $service->save();
                 $status = $this->_dictionaries->getStatusList('orders')->find('invoiced', 'acronym');
+                $statusReleased = $this->_dictionaries->getStatusList('orders')->find('released', 'acronym');
+                $this->_orderlines->setLazyLoading(true);
                 foreach ($service->getProducts() as $product) {
+                    if ($orderLine = $this->_orderlines->find($product->productid)->current()) {
+                        $orderLine->statusid = $statusReleased->id;
+                        $orderLine->serviceid = null;
+                        $orderLine->qtyavailable += $orderLine->quantity;
+                        if ($orderLine->quantity > $orderLine->qtyavailable) {
+                            $form->getElement('productid')->setErrors(array('productid' => 'Wystąpił problem z dodaniem produktu'));
+                            return;
+                        }
+                        $orderLine->save();
+                    }
                     $product->delete();
                 }
                 $productIds = array_filter($request->getParam('productid'));
@@ -1630,6 +1699,14 @@ class Services_ServicesController extends Application_Controller_Abstract {
                         } else {
                             $params['productid'] = $product->id;
                             $params['productname'] = $product->getProduct()->name;
+                            $orderLine = $this->_orderlines->find($orderLineId)->current();
+                            $orderLine->statusid = $status->id;
+                            $orderLine->qtyavailable -= $orderLine->quantity;
+                            if ($orderLine->qtyavailable < 0) {
+                                $form->getElement('productid')->setErrors(array('productid' => 'Wystąpił problem z dodaniem produktu'));
+                                return;
+                            }
+                            $orderLine->save();
                         }
                         $serviceProduct = $table->createRow($params);
                         //$serviceProduct->serviceid = $service->id;
@@ -1638,7 +1715,7 @@ class Services_ServicesController extends Application_Controller_Abstract {
                             $serviceProduct->save();
                             
                         //}catch(Exception $e){var_dump($orderLineId,$e->getMessage(),$serviceProduct->toArray());exit;}
-                        $this->_orderlines->update(array('statusid' => $status->id, 'serviceid' => $service->id), $this->_orderlines->getAdapter()->quoteInto('id = ?', $orderLineId));
+                        //$this->_orderlines->update(array('statusid' => $status->id, 'serviceid' => $service->id), $this->_orderlines->getAdapter()->quoteInto('id = ?', $orderLineId));
                     }
                 }
                 foreach ($service->getCodes() as $attribute) {
@@ -1877,7 +1954,18 @@ class Services_ServicesController extends Application_Controller_Abstract {
                     $service->performed = null;
                     $service->technicalcomments = null;
                     $service->save();
+                    $statusReleased = $this->_dictionaries->getStatusList('orders')->find('released', 'acronym');
+                    $this->_orderlines->setLazyLoading(true);
                     foreach ($service->getProducts() as $product) {
+                        $orderLine = $this->_orderlines->find($product->productid)->current();
+                        $orderLine->statusid = $statusReleased->id;
+                        $orderLine->serviceid = null;
+                        $orderLine->qtyavailable += $orderLine->quantity;
+                        if ($orderLine->quantity > $orderLine->qtyavailable) {
+                            $form->getElement('id-' . $i)->setErrors(array('productid' => 'Wystąpił problem z dodaniem produktu'));
+                            return;
+                        }
+                        $orderLine->save();
                         $product->delete();
                     }
                     foreach ($service->getCodes() as $attribute) {
